Move PageSEO logic to Layout
