--- src/components/page-seo/index.ts
@@ -1,3 +1,0 @@
-import { PageSEO } from "./page-seo";
-
-export default PageSEO;
