+++ src/pages/contact.tsx
@@ -4,14 +4,12 @@ import { NextPage } from "next";
 import { IPageProps } from "../interfaces/models";
 import { ReactLocaleContext } from "../contexts/locale-context";
 import ContactForm from "../components/contact-form";
-import PageSEO from "../components/page-seo";
 import { classes } from "./page-base.st.css";
 
 const Contact: NextPage<IPageProps> = () => {
 	const { translate } = useContext(ReactLocaleContext);
 	return (
 		<Layout>
-			<PageSEO />
 			<article className={classes.root}>
 				<h1 className={classes.heading3}>{translate("CONTACT_PAGE_TITLE")}</h1>
 				<h2 className={classes.title}>{translate("CONTACT_PAGE_SUBTITLE")}</h2>
